New `~runhere` command
The `~runhere` command will do the same thing as `~run`, in that it runs a shell command or a shell builtin.
However, unlike `~run`, it'll operate on the buffer contents themselves, writing a temporary file and running the provided command on the provided file.
This can be useful for running Python or another such interpreter on your code without having to save the file first.