Fixed sued just generally not working
Writing to the buffer now works as it did before, instead of just writing the first word of the typed in sentence.
`process_command`  now returns `ExitStatus`, which can be either `Success` or `Failure` - if it returns `ExitStatus::Failure`, sued will break the `ReadResult::Input(line) = interface.read_line().unwrap()` loop and exit.